色々修正 & 追加
