+++ src/main/java/mumei/moguratataki/game/GameControl.java
@@ -5,7 +5,9 @@ import mumei.moguratataki.game.event.GameEndEvent;
 import mumei.moguratataki.game.event.GameStartEvent;
 import mumei.moguratataki.game.event.GameUpdateEvent;
 import mumei.moguratataki.game.event.PreGameUpdateEvent;
+import mumei.moguratataki.listener.*;
 import mumei.moguratataki.utility.Timer;
+import mumei.moguratataki.utility.Util;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.title.Title;
 import org.bukkit.Bukkit;
@@ -22,15 +24,24 @@ public class GameControl {
         if (gameTime < 0) throw new IllegalArgumentException("ゲーム時間は0秒以上にする必要があるよ");
         this.preGameTime = preGameTime;
         this.gameTime = gameTime;
+
+        Util.registerEvents(
+                Moguratataki.getplugin(),
+                new GameEndListener(),
+                new GameListener(),
+                new GameUIListener(),
+                new MoguraListener(),
+                new PlayerListener(),
+                new PreGameUIListener(),
+                new SpectatorListener()
+        );
     }
 
     public void start() {
         if (started) stop();
         preGameTimer = new Timer(preGameTime);
         preGameTimer.setOnUpdate(() -> {
-            Bukkit.getPluginManager().callEvent(new PreGameUpdateEvent());
-            Bukkit.broadcast(Component.text(preGameTimer.getCurrentTime() + "秒前"));
-            Bukkit.getServer().showTitle(Title.title(Component.text(""), Component.text(preGameTimer.getCurrentTime() + "秒前")));
+            Bukkit.getPluginManager().callEvent(new PreGameUpdateEvent(gameTimer.getCurrentTime()));
         });
         preGameTimer.setOnEnd(() -> {
             gameTimer.start();
@@ -42,13 +53,12 @@ public class GameControl {
 
         gameTimer = new Timer(gameTime);
         gameTimer.setOnUpdate(() -> {
-            Bukkit.getPluginManager().callEvent(new GameUpdateEvent());
-            Bukkit.getServer().sendActionBar(Component.text("⌚ " + gameTimer.getCurrentTime()));
+            Bukkit.getPluginManager().callEvent(new GameUpdateEvent(gameTimer.getCurrentTime()));
         });
         gameTimer.setOnEnd(() -> {
-            Bukkit.getPluginManager().callEvent(new GameEndEvent(Moguratataki.MoguratatakiTeam.MOGURA.getTeam()));
             fixedStarted = false;
             started = false;
+            Bukkit.getPluginManager().callEvent(new GameEndEvent(Moguratataki.MoguratatakiTeam.MOGURA.getTeam()));
         });
 
         started = true;
