+++ src/main/java/mumei/moguratataki/game/event/GameUpdateEvent.java
@@ -5,12 +5,18 @@ import org.bukkit.event.HandlerList;
 
 public class GameUpdateEvent extends Event {
     private static final HandlerList HANDLERS = new HandlerList();
+    private final int time;
 
-    public GameUpdateEvent() {
+    public GameUpdateEvent(int time) {
+        this.time = time;
     }
 
     public static HandlerList getHandlerList() {
         return HANDLERS;
+    }
+
+    public int getTime() {
+        return time;
     }
 
     @Override
