+++ src/main/java/mumei/moguratataki/listener/GameEndListener.java
@@ -1,10 +1,13 @@
 package mumei.moguratataki.listener;
 
+import mumei.moguratataki.Moguratataki;
 import mumei.moguratataki.game.event.GameEndEvent;
 import mumei.moguratataki.team.Team;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.format.TextColor;
 import org.bukkit.Bukkit;
+import org.bukkit.GameMode;
+import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 
@@ -20,6 +23,11 @@ public final class GameEndListener implements Listener {
         } else {
             // どちらかが勝利の時
             Bukkit.broadcast(Component.text("勝ったチーム: ").append(winnerTeam.getTeam().displayName()).color(TextColor.color(226, 85, 74)));
+        }
+
+        for (Player player : Bukkit.getOnlinePlayers()) {
+            Moguratataki.MoguratatakiTeam.SPEC.getTeam().addPlayer(player);
+            player.setGameMode(GameMode.ADVENTURE);
         }
     }
 }
